\name{maize.Pheno}
\alias{maize.Pheno}
\title{Phenotypical data for the maize dataset}
\description{
Phenotypic data for grain yield and plant height for 100 out of 
400 possible crosses originated from 40 inbreed lines belonging 
to two heterotic groups, 20 lines in each. The data are stored in
a data.frame with 6 columns: Location, GCA1 (Parent 1), CGA2 (Parent 2), 
SCA (hybrid), Yield and PlantHeigh. Records with missing values in 
the last two columns corresponds to hybrids that were not evaluated in 
the field and that we need to predict.
}

\keyword{datasets}
